Add session interface
Allow cache management functions to be invoked using an alternative
syntax which is friendly to type checking.  This is done by passing
a special session argument.

Also allow users to specify extra functions which will skip
invalidation.  Again this allows for an alternative method than getval
and putval which are friendly to type checking.

Signed-off-by: Isaac To <isaacto@gmail.com>