Fix misleading stack trace of cached exception
After an exception is raised, it will still be modified outside the smemo
system by adding frames to it.  So when a cached exception is raised, it
will carry the old frames plus some new ones, which becomes old ones upon
the next iteration.  The resulting exception will usually contain many
frames collected over previous tries.  We now empty the stack trace, which
is more reasonable because we really didn't have called the inside of the
function.

Signed-off-by: Isaac To <isaacto@gmail.com>